requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
et_xmlfile==2.0.0
filelock==3.17.0
fsspec==2025.2.0
huggingface-hub==0.29.2
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
psutil==7.0.0
python-dotenv==1.0.1
python-telegram-bot==13.7
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
schedule==1.2.2
six==1.17.0
sympy==1.13.1
tokenizers==0.21.0
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
